requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
djangorestframework==3.7.7
honcho==1.0.1
